+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [UIMARCAUTH-123](https://issues.folio.org/browse/UIMARCAUTH-123) Fix losing third pane after search and editing first result.
 - [UIMARCAUTH-125](https://issues.folio.org/browse/UIMARCAUTH-125) Replace `babel-eslint` with `@babel/eslint-parser`.
 - [UIMARCAUTH-120](https://issues.folio.org/browse/UIMARCAUTH-120) Allow user to select MARC authority record(s) to export
+- [UIMARCAUTH-130](https://issues.folio.org/browse/UIMARCAUTH-130) Add a Select all/Un-select all on Results List
 
 ## [1.0.5](https://github.com/folio-org/ui-marc-authorities/tree/v1.0.5) (2022-04-15)